Adds authorization to all axios rest calls.
